+++ benchmark/benches/bench_query_posvel_soa_unpack.jl
@@ -1,8 +1,8 @@
 
 function setup_query_posvel_soa_unpack(n_entities::Int)
     world = World(
-        Position => StructArrayStorage,
-        Velocity => StructArrayStorage,
+        Position => Storage{StructArray},
+        Velocity => Storage{StructArray},
     )
     for i in 1:n_entities
         new_entity!(world, (Position(i, i * 2), Velocity(1, 1)))