+++ src/query.jl
@@ -289,13 +289,13 @@ end
         push!(exprs, :(@inbounds $col_sym = $stor_sym.data[table.id]))
 
         if is_optional[i] === Val{true}
-            if storage_modes[i] != StructArrayStorage && fieldcount(comp_types[i]) > 0
+            if storage_modes[i] != Storage{StructArray} && fieldcount(comp_types[i]) > 0
                 push!(exprs, :($vec_sym = length($col_sym) == 0 ? nothing : FieldViewable($col_sym)))
             else
                 push!(exprs, :($vec_sym = length($col_sym) == 0 ? nothing : view($col_sym, :)))
             end
         else
-            if storage_modes[i] != StructArrayStorage && fieldcount(comp_types[i]) > 0
+            if storage_modes[i] != Storage{StructArray} && fieldcount(comp_types[i]) > 0
                 push!(exprs, :($vec_sym = FieldViewable($col_sym)))
             else
                 push!(exprs, :($vec_sym = view($col_sym, :)))
@@ -333,10 +333,10 @@ Base.IteratorSize(::Type{<:Query}) = Base.SizeUnknown()
     for i in 1:N
         T = comp_types[i]
 
-        ST = :(storage_type($(storage_modes[i]), $T))
+        ST = :(_storage_type($(storage_modes[i]), $T))
         base_view = if fieldcount(comp_types[i]) == 0
             :(SubArray{$T,1,$ST,Tuple{Base.Slice{Base.OneTo{Int}}},IndexStyle($ST)==IndexLinear()})
-        elseif storage_modes[i] != StructArrayStorage
+        elseif storage_modes[i] != Storage{StructArray}
             :(_FieldsViewable_type($ST))
         else
             :(_StructArrayView_type($T, UnitRange{Int}))