+++ src/storage.jl
@@ -3,27 +3,19 @@ struct _ComponentStorage{C,A<:AbstractArray{C,1}}
     data::Vector{A}
 end
 
-"""
-    new_storage(ModeType, ComponentType)
-
-Returns a new empty instance of the storage type.
-"""
-function new_storage(::Type{S}, ::Type{C}) where {S<:AbstractStorage,C}
-    storage_type(S, C)()
-end
-function new_storage(::Type{StructArrayStorage}, ::Type{C}) where {C}
-    _StructArray(C)
+function _new_storage(::Type{S}, ::Type{C}) where {S<:Storage,C}
+    _storage_type(S, C)()
 end
 
-"""
-    storage_type(ModeType, ComponentType)
-
-Returns the storage type of the input mode.
-"""
-function storage_type(::Type{VectorStorage}, ::Type{C}) where {C}
-    Vector{C}
+function _new_storage(::Type{Storage{StructArray}}, ::Type{C}) where {C}
+    StructArray(C)
 end
-function storage_type(::Type{StructArrayStorage}, ::Type{C}) where {C}
+
+function _storage_type(::Type{<:Storage{T}}, ::Type{C}) where {T,C}
+    T{C}
+end
+
+function _storage_type(::Type{Storage{StructArray}}, ::Type{C}) where {C}
     _StructArray_type(C)
 end
 
@@ -44,9 +36,9 @@ function _set_component!(s::_ComponentStorage{C,A}, arch::UInt32, row::UInt32, v
 end
 
 @generated function _add_column!(storage::_ComponentStorage{C,A}) where {C,A<:AbstractArray}
-    if A <: _StructArray
+    if A <: StructArray
         return quote
-            push!(storage.data, _StructArray(C))
+            push!(storage.data, StructArray(C))
         end
     else
         return quote
@@ -88,7 +80,7 @@ end
     old_table::UInt32,
     new_table::UInt32,
     row::UInt32,
-) where {C,A<:_StructArray}
+) where {C,A<:StructArray}
     names = fieldnames(A.parameters[1])
     exprs_push_remove = Expr[]
     for name in names
@@ -158,7 +150,7 @@ end
     s::_ComponentStorage{C,A},
     arch::UInt32,
     row::UInt32,
-) where {C,A<:_StructArray}
+) where {C,A<:StructArray}
     names = fieldnames(A.parameters[1])
     exprs_remove = Expr[]
     for name in names