+++ test/setup_default.jl
@@ -2,4 +2,4 @@
 const N_fake = 0
 const offset_ID = 0
 const M_mask = 1
-const DefaultStorage = VectorStorage
+const DefaultStorage = Storage{Vector}