+++ test/test_world.jl
@@ -11,7 +11,7 @@ end
 @testset "World creation 2" begin
     world = World(
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
         Altitude,
         ChildOf,
     )
@@ -24,12 +24,12 @@ end
         "ArgumentError: Component type Health not found in the World",
         _component_id(params, Health))
 
-    @test isa(_get_storage(world, Position), _ComponentStorage{Position,storage_type(DefaultStorage, Position)})
-    @test isa(_get_storage(world, Position).data[1], storage_type(DefaultStorage, Position))
+    @test isa(_get_storage(world, Position), _ComponentStorage{Position,_storage_type(DefaultStorage, Position)})
+    @test isa(_get_storage(world, Position).data[1], _storage_type(DefaultStorage, Position))
     @test isa(_get_storage(world, Velocity), _ComponentStorage{Velocity,_StructArray_type(Velocity)})
     @test isa(_get_storage(world, Velocity).data[1], _StructArray_type(Velocity))
-    @test isa(_get_storage(world, Altitude), _ComponentStorage{Altitude,storage_type(DefaultStorage, Altitude)})
-    @test isa(_get_storage(world, Altitude).data[1], storage_type(DefaultStorage, Altitude))
+    @test isa(_get_storage(world, Altitude), _ComponentStorage{Altitude,_storage_type(DefaultStorage, Altitude)})
+    @test isa(_get_storage(world, Altitude).data[1], _storage_type(DefaultStorage, Altitude))
 
     @test length(_get_relations_storage(world, Position).archetypes) == 0
     @test length(_get_relations_storage(world, Position).targets) == 0
@@ -60,10 +60,10 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
     )
 
-    @test isa(_get_storage(world, Position), _ComponentStorage{Position,storage_type(DefaultStorage, Position)})
+    @test isa(_get_storage(world, Position), _ComponentStorage{Position,_storage_type(DefaultStorage, Position)})
     @test isa(_get_storage(world, Velocity), _ComponentStorage{Velocity,_StructArray_type(Velocity)})
 end
 
@@ -73,7 +73,7 @@ end
         # TODO: type instability here. Add benchmarks for world creation.
         @test_opt World(
             Position,
-            Velocity => StructArrayStorage,
+            Velocity => Storage{StructArray},
         )
     end
 end
@@ -89,16 +89,16 @@ end
         World(Position, Velocity, Relationship))
 
     @test_throws(
-        "ArgumentError: Health is not a valid storage mode, must be a subtype of AbstractStorage",
+        "ArgumentError: Health is not a valid storage mode, must be Storage{T<:AbstractVector}",
         World(Position, Velocity, Altitude => Health))
 
     @test_throws(
-        "ArgumentError: can't use StructArrayStorage for Int64 because it has no fields",
-        World(Int64 => StructArrayStorage))
+        "ArgumentError: can't use Storage{StructArray} for Int64 because it has no fields",
+        World(Int64 => Storage{StructArray}))
 
     @test_throws(
-        "ArgumentError: can't use StructArrayStorage for LabelComponent because it has no fields",
-        World(LabelComponent => StructArrayStorage))
+        "ArgumentError: can't use Storage{StructArray} for LabelComponent because it has no fields",
+        World(LabelComponent => Storage{StructArray}))
 end
 
 @testset "World creation large" begin
@@ -179,7 +179,7 @@ end
     @test isa(id_int, Int)
     @test world._registry.types[id_int] == Int
     @test length(world._storages) == N_fake + 2
-    @test world._storages[id_int] isa _ComponentStorage{Int,storage_type(DefaultStorage,Int)}
+    @test world._storages[id_int] isa _ComponentStorage{Int,_storage_type(DefaultStorage,Int)}
     @test length(world._storages[id_int].data) == 1
 
     # Register Position component
@@ -187,7 +187,7 @@ end
     @test isa(id_pos, Int)
     @test world._registry.types[id_pos] == Position
     @test length(world._storages) == N_fake + 2
-    @test world._storages[id_pos] isa _ComponentStorage{Position,storage_type(DefaultStorage,Position)}
+    @test world._storages[id_pos] isa _ComponentStorage{Position,_storage_type(DefaultStorage,Position)}
     @test length(world._storages[id_pos].data) == 1
 
     # Re-register Int component (should not add new storage)
@@ -204,7 +204,7 @@ end
     _ = World(Position, MutableComponent; allow_mutable=true)
 
     @test_throws("ArgumentError: Component type MutableComponent must be immutable because it uses StructArray storage",
-        World(Position, MutableComponent => StructArrayStorage))
+        World(Position, MutableComponent => Storage{StructArray}))
 end
 
 @testset "_get_storage Tests" begin
@@ -212,11 +212,11 @@ end
     params = typeof(world).parameters[1]
 
     storage1 = _get_storage(world, Int)
-    @test storage1 isa _ComponentStorage{Int,storage_type(DefaultStorage, Int)}
+    @test storage1 isa _ComponentStorage{Int,_storage_type(DefaultStorage, Int)}
 
     id = _component_id(params, Int)
     storage2 = _get_storage(world, Int)
-    @test storage2 isa _ComponentStorage{Int,storage_type(DefaultStorage, Int)}
+    @test storage2 isa _ComponentStorage{Int,_storage_type(DefaultStorage, Int)}
 
     @test storage1 === storage2
 
@@ -294,8 +294,8 @@ end
     pos_storage = _get_storage(world, Position)
     vel_storage = _get_storage(world, Velocity)
 
-    @test isa(pos_storage, _ComponentStorage{Position,storage_type(DefaultStorage, Position)})
-    @test isa(vel_storage, _ComponentStorage{Velocity,storage_type(DefaultStorage, Velocity)})
+    @test isa(pos_storage, _ComponentStorage{Position,_storage_type(DefaultStorage, Position)})
+    @test isa(vel_storage, _ComponentStorage{Velocity,_storage_type(DefaultStorage, Velocity)})
     @test length(pos_storage.data) == 3
     @test length(vel_storage.data) == 3
 end
@@ -339,7 +339,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
     )
 
     e1 = new_entity!(world, (Position(1, 2), Velocity(3, 4)))
@@ -370,7 +370,7 @@ end
     @testset "World get/set components JET" begin
         world = World(
             Position,
-            Velocity => StructArrayStorage,
+            Velocity => Storage{StructArray},
         )
         e1 = new_entity!(world, (Position(1, 2), Velocity(3, 4)))
 
@@ -383,7 +383,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
     )
 
     entity = new_entity!(world, ())
@@ -413,7 +413,7 @@ end
     @testset "World new_entity! JET" begin
         world = World(
             Position,
-            Velocity => StructArrayStorage,
+            Velocity => Storage{StructArray},
         )
 
         using FunctionWrappers
@@ -433,7 +433,7 @@ end
         Dummy,
         Position,
         ChildOf,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
     )
 
     parent1 = new_entity!(world, ())
@@ -664,7 +664,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
         ChildOf,
     )
 
@@ -706,7 +706,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
         Altitude,
         ChildOf,
     )
@@ -742,7 +742,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
         NoIsBits,
         MutableComponent,
         MutableNoIsBits;
@@ -795,7 +795,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
         Altitude,
     )
 
@@ -842,7 +842,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
         Altitude,
     )
 
@@ -862,7 +862,7 @@ end
         @test length(pos_col) == 100
         @test length(vel_col) == 100
         @test pos_col isa FieldViewable
-        @test vel_col isa StructArrayView
+        @test vel_col isa _StructArrayView
         for i in eachindex(ent)
             @test is_alive(world, ent[i]) == true
             @test pos_col[i] == Position(99, 99)
@@ -940,7 +940,7 @@ end
     @testset "World new_entities! JET" begin
         world = World(
             Position,
-            Velocity => StructArrayStorage,
+            Velocity => Storage{StructArray},
         )
         using FunctionWrappers
         excluded = Set([
@@ -959,7 +959,7 @@ end
     world = World(
         Dummy,
         Position,
-        Velocity => StructArrayStorage,
+        Velocity => Storage{StructArray},
         Altitude,
         Health,
     )
@@ -1080,7 +1080,7 @@ end
 @testset "World add/remove components batch" begin
     world = World(
         Dummy,
-        Position => StructArrayStorage,
+        Position => Storage{StructArray},
         Velocity,
         Altitude,
         Health,
@@ -1126,7 +1126,7 @@ end
         world = World(
             Dummy,
             Position,
-            Velocity => StructArrayStorage,
+            Velocity => Storage{StructArray},
         )
         using FunctionWrappers
         excluded = Set([
@@ -1181,10 +1181,10 @@ end
 @testset "World exchange components batch" begin
     world = World(
         Dummy,
-        Position => StructArrayStorage,
+        Position => Storage{StructArray},
         Velocity,
         Altitude,
-        Health => StructArrayStorage,
+        Health => Storage{StructArray},
     )
 
     new_entities!(world, 10, (Position(1, 1), Altitude(100)))